test: migrate zero-integrations-slack-upload-complete to bdd
